change the probability distribution for valence quark positions
The program originally used a Woods-Saxon distribution for the valence
quark. The parameter are quesionable, especially we found the v2 in pA
collisions were not enough compared to the experimental data. Now we
change the probability distribution of the valence quarks to a 3D Gaussian
distribution whose parameters are determined by the fits to HERA data.