introduce two variance for the yLoss parameterization
Based on fitting to the experimentally measured dN^ch/deta at the RHIC
and LHC energies, it seems that the variance of the y_loss distributions
increases from the top RHIC energy to the LHC energies. So now we
introduce two parameters in the parameter files to specify the variance
at the top RHIC energy and LHC energies.