Fixed a Nan error
Only add a soft parton contribution when the leftover energy is larger
than the valence quark mass. There were some crashes of the simulations
because the E < M_valence, especially at low collision energies. In
these case, the rapidity computed from arccosh(E/M) will return Nan and
triggers an error in the assert function. So we will only add a soft
parton contribution when the leftover energy of the nucleon is larger
than the M_valence quark. This treatment is actually more physical as
the soft parton will play a role at high-energy collisions.