set the remnant strings to 2 units of rapidity
At the LHC energy, the remnant string length (yrem/2, yrem) seems to be
too long such that the centrality cut from the V0A detector
\eta \in (2.8, 5.1) was affected by the remnant strings. Now, we set the
remnant string length to 2 units of rapidity. For 5.02 TeV, this
approach leads to remnant strings from (6, 8) in rapidity. In this case,
the previous tune 14 should work.