improved the sampling the pdf using inverse CDF
The cut-off artifacts is because the CDF function increases very slow
near x = 0 and x = 1. So a linear spacing in the inverse CDF behaves
poorly near these two regions. Therefore, a small dx~1e-4 is required to
generate good samples around x = 0 and x = 1. The solution is to
directly use CDF and perform the inversion on the fly. In this way, we
can use dx = 0.01 to get same performance as dx = 1e-4.

With this new implementation, the effeciency in the Metropolis algorithm
seems to reduce more. I tested the samples after Metropolis steps and
find that 1% violation in the sample does not generate visible change in
the single x pdf. So we set the allowed violation to be 0.5% in the code
to ensure a good sample of the x and the code can finish within a few
minutes.