Restructure the get_nucleus_density function
We add a generic function to compute the nucleus density. Then this
function is called to get local nuclear density for the projectile and
target nuclei.

The total nuclear density can be get by summing over the projectile and
target nuclear densities.

We also did some clean up of the code.